requirements.txt
aiohttp==3.11.12
curl_cffi==0.9.0
httpx==0.28.1
loguru==0.7.3
openai==1.65.4
openpyxl==3.1.5
prompt_toolkit==3.0.50
PyYAML==6.0.2
rich==13.9.4
tabulate==0.9.0
urllib3==2.3.0
google-generativeai==0.7.1
